Updated some stylings, moved the about page into the homepage for users who are not signed in.
