+++ src/app/components/HowItWorks.tsx
