+++ src/middleware.ts
@@ -12,5 +12,5 @@ export function middleware(request: NextRequest) {
 }
 
 export const config = {
-  matcher: ['/', '/invite/:path*'],
+  matcher: ['/invite/:path*'],
 };
